// packages/engine/src/templates/character/CharacterStateTypes.d.ts
export declare enum CharacterStateTypes {
    IDLE = 0,
    MOVING = 1,
    JUMP_IDLE = 2,
    JUMP_RUNNING = 3,
    FALLING = 4,
    CROUCHING = 5,
    WALK = 6,
    SPRINT = 7,
    INTERACTING = 8,
    MOVING_FORWARD = 9,
    MOVING_BACKWARD = 10,
    MOVING_LEFT = 11,
    MOVING_RIGHT = 12,
    IDLE_ROTATE_LEFT = 13,
    IDLE_ROTATE_RIGHT = 14,
    WALK_END = 15,
    WALK_START_BACKWARD = 16,
    WALK_START_FORWARD = 17,
    WALK_START_LEFT = 18,
    WALK_START_RIGHT = 19,
    SPRINT_LEFT = 20,
    SPRINT_RIGHT = 21,
    SPRINT_BACKWARD = 22,
    DROP_IDLE = 23,
    DROP_ROLLING = 24,
    DROP_RUNNING = 25,
    DRIVING_IDLE = 26,
    ENTER_VEHICLE = 27,
    EXIT_VEHICLE = 28
}